+++ doomsday/plugins/common/src/p_map.c
@@ -2345,7 +2345,7 @@ void P_UseLines(player_t* player)
     {
 #ifdef _DEBUG
         Con_Message("P_UseLines: Sending a use request for player %i.\n",
-                    player - players);
+                    (int)(player - players));
 #endif
         NetCl_PlayerActionRequest(player, GPA_USE, 0);
         return;
